refactor(sprint4): split config_base.py into 6 specialized modules
Decomposes 1155 LOC god config into data_repo, ai_providers, intervals,
week_reference, email_config, withings_config. config_base.py retained
as re-export shim for backward compatibility. Also adds post-rebase
workflow and pydantic-core fix to CONTRIBUTING.md.
